refactor: extract ui panel rendering
