+++ AGENTS.md
@@ -34,6 +34,7 @@ cargo test --features test-utils                         # Run all tests includi
 | `src/managers/common.rs`      | `DeviceManager` trait                                                                    |
 | `src/ui/render.rs`            | Three-panel layout rendering and panel composition                                       |
 | `src/ui/dialogs.rs`           | Modal dialog rendering (`create/confirm/api-level/notifications`)                        |
+| `src/ui/panels.rs`            | Panel rendering helpers (`android/ios/details/logs/commands`)                            |
 | `src/constants/`              | All constants (NO hardcoded values in source)                                            |
 | `src/models/`                 | Core data structures                                                                     |
